fixes #7
